recording/requirements.txt
numpy==1.13.3
matplotlib==2.1.0
pyserial==3.4
